server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "senks",email = "yungync@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn (>=0.34.1,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
